OP-151 Staking contracts (#92)
* Basis of OPTY Token distribution

* Implemented calculations in OptyMinter.sol. Also added modifier onlyMinter and updated OPTY.sol contract

* Modified functions to calculate amount of OPTY accrued to avoid rounding errors

* Updated basic pool contract and OPTYMinter.sol. Working e2e

* del contracts/Registry, resolve OPTYMinter.sol

* Updated ModifiersController.sol

* Updated contracts to support pool operations without rebalance

* Included lastUserUpdate mapping. OPTY Distribution working e2e for interactions with no rebalance.

* Updated all pool contracts

* Updated OPTYMinter.sol. currentOptyPoolIndex function added

* Updated .gitignore

* Removed desktop.ini and package-lock.json

* Staking contracts created

* Pool.sol and OPTYMinter.sol updated to support staking

* Fixed updatePool function in OPTYStakingPool.sol

* Removed UserState structure (we only need to store last interaction timestamp of each user)

* Added getBlockTimestamp() function to OPTYStakingPool.sol contract

* Fixed bug in Pool.sol (functions that interact with OPTYMinter were duplicated)

* Modified function in OPTYMinter.sol to distribute rewards per second instead of per block

* apply lint

+++ contracts/pools/advanced-plus/Pool.sol
@@ -191,7 +191,7 @@ contract AdvancePlusPool is ERC20, ERC20Detailed, Modifiers, ReentrancyGuard, Po
     }
 
     function _batchMintAndBurn() internal returns (bool _success) {
-        uint iterator = first;
+        uint256 iterator = first;
         while (last >= iterator) {
             optyMinterContract.updateSupplierRewards(address(this), queue[iterator].account);
             if (queue[iterator].isDeposit) {
@@ -212,7 +212,7 @@ contract AdvancePlusPool is ERC20, ERC20Detailed, Modifiers, ReentrancyGuard, Po
             delete queue[first];
             first++;
         }
-        
+
         _success = true;
     }
 
@@ -304,7 +304,7 @@ contract AdvancePlusPool is ERC20, ERC20Detailed, Modifiers, ReentrancyGuard, Po
             _withdrawAll();
             harvest(strategyHash);
         }
-        
+
         optyMinterContract.updateSupplierRewards(address(this), msg.sender);
         // subtract pending deposit from total balance
         _redeemAndBurn(msg.sender, balance().sub(depositQueue), _redeemAmount);
